+++ src/components/Mypage/Orderlist.js
@@ -1,5 +1,5 @@
 import orderlistCss from './Orderlist.module.scss';
-import css from './Mypage.module.scss';
+import css from '../../pages/Mypage/Mypage.module.scss';
 import { useEffect, useState } from 'react';
 
 function Orderlist() {
