+++ src/components/Mypage/Point.module.scss
