--- src/pages/Mypage/Profile.js
@@ -1,7 +1,0 @@
-import css from './Profile.module.scss';
-
-function Profile() {
-  return <div>안녕하세요 내 정보 입니다.</div>;
-}
-
-export default Profile;
