+++ src/styles/reset.scss
@@ -8,6 +8,11 @@
   box-sizing: border-box;
 }
 
+body {
+  height: 100vh;
+  background: #eee;
+}
+
 ul,
 ol {
   list-style: none;
